--- a/core/network-libp2p/src/behaviour.rs
+++ b/core/network-libp2p/src/behaviour.rs
@@ -145,9 +145,20 @@ impl<TSubstream> NetworkBehaviourEventProcess<IdentifyEvent> for Behaviour<TSubs
 }
 
 impl<TSubstream> NetworkBehaviourEventProcess<KademliaOut> for Behaviour<TSubstream> {
-	fn inject_event(&mut self, _: KademliaOut) {
+	fn inject_event(&mut self, out: KademliaOut) {
 		// We only ever use Kademlia for discovering nodes, and nodes discovered by Kademlia are
 		// automatically added to the topology. Therefore we don't need to do anything.
+		match out {
+			KademliaOut::FindNodeResult { key, closer_peers } => {
+				trace!(target: "sub-libp2p", "Kademlia query for {:?} yielded {:?} results",
+					key, closer_peers.len());
+				if closer_peers.is_empty() {
+					warn!(target: "sub-libp2p", "Kademlia random query has yielded empty results");
+				}
+			}
+			// We never start any GET_PROVIDERS query.
+			KademliaOut::GetProvidersResult { .. } => ()
+		}
 	}
 }