Improves peers discovery and the encryption handshake performances.
Integrates https://github.com/libp2p/rust-libp2p/pull/855 and https://github.com/libp2p/rust-libp2p/pull/856

Running the tests in `node/cli` yields something like:

Before:
```
real    0m48,839s
user    0m9,830s
sys     0m2,104s
```
After:
```
real    0m8,442s
user    0m2,685s
sys     0m0,988s
```

cc @arkpar who reported the problem